@@ -4,11 +4,10 @@
 #
 #  SPDX-License-Identifier: LGPL-3.0-or-later
 #
-from typing import List
 
 
 class Basket:
-    items: List[str] = []
+    items: list[str] = []
 
     def add_item(self, name: str):
         self.items.append(name)
