@@ -4,32 +4,31 @@
 #
 #  SPDX-License-Identifier: LGPL-3.0-or-later
 #
-from typing import Dict, List, Set, Tuple, Union
 
 
-def nested_list(x: List[List[str]]):
+def nested_list(x: list[list[str]]):
     for e in x:
         for i in e:
             print(i)
 
 
-def nested_no_param(x: List[Set]):
+def nested_no_param(x: list[set]):
     for e in x:
         for i in e:
             print(i)
 
 
-def list_union(x: List[Union[int, str]]):
+def list_union(x: list[int | str]):
     for e in x:
         print(e)
 
 
-def check_tuple(x: Tuple[int, str]):
+def check_tuple(x: tuple[int, str]):
     if x[0] == 5:
         print("foo")
 
 
-def untyped_tuple(x: Tuple):
+def untyped_tuple(x: tuple):
     if len(x) == 3:
         print("foo")
 
@@ -39,6 +38,6 @@ def untyped_list(x):
         print("foo")
 
 
-def dict_str_int(dic: Dict[str, int]):
+def dict_str_int(dic: dict[str, int]):
     for x, y in dic.items():
         print(f"{x}: {y+1}")
