@@ -5,7 +5,6 @@
 #  SPDX-License-Identifier: LGPL-3.0-or-later
 #
 import array
-from typing import Optional
 
 
 class Queue:
@@ -33,7 +32,7 @@ class Queue:
             self.tail = 0
         return True
 
-    def dequeue(self) -> Optional[int]:
+    def dequeue(self) -> int | None:
         if self.size == 0:
             return None
         x = self.data[self.head]
