@@ -27,7 +27,7 @@ def seeding():
 @pytest.fixture
 def fixture_dir():
     return os.path.join(
-        os.path.dirname(os.path.abspath(__file__)),
+        os.path.dirname(__file__),
         "",
         "..",
         "fixtures",
